requirements.txt
matplotlib==3.9.1
numpy==2.0.1
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
scikit_learn==1.5.1
spotipy==2.24.0
streamlit==1.36.0
streamlit_option_menu==0.3.13
